+++ pliers/filters/image.py
@@ -43,6 +43,69 @@ class ImageCroppingFilter(ImageFilter):
                          data=new_img)
 
 
+class ImageResizingFilter(ImageFilter):
+
+    ''' Resizes an image, while optionally maintaining aspect ratio.
+
+    Args:
+        size (tuple of two ints): new size of the image.
+        maintain_aspect_ratio (boolean): if true, resize the image while
+            maintaining aspect ratio, and pad the rest with zero values.
+            Otherwise, potentially distort the image during resizing to fit the
+            new size.
+        resample str: resampling method. One of 'nearest', 'bilinear',
+            'bicubic', 'lanczos', 'box', and 'hamming'. See
+            https://pillow.readthedocs.io/en/5.1.x/handbook/concepts.html#concept-filters
+            for more information.
+    '''
+
+    _log_attributes = ('size', 'maintain_aspect_ratio', 'resample')
+    VERSION = '1.0'
+
+    def __init__(self, size, maintain_aspect_ratio=False, resample='bicubic'):
+        self.size = size
+        self.maintain_aspect_ratio = maintain_aspect_ratio
+        resampling_mapping = {
+            'nearest': Image.NEAREST,
+            'bilinear': Image.BILINEAR,
+            'bicubic': Image.BICUBIC,
+            'lanczos': Image.LANCZOS,
+            'box': Image.BOX,
+            'hamming': Image.HAMMING,
+        }
+        if resample.lower() not in resampling_mapping.keys():
+            raise ValueError(
+                "Unknown resampling method '{}'. Allowed values are '{}'"
+                .format(resample, "', '".join(resampling_mapping.keys())))
+        self.resample = resampling_mapping[resample]
+        super(ImageResizingFilter, self).__init__()
+
+    def _filter(self, stim):
+        pillow_img = Image.fromarray(stim.data)
+
+        if not self.maintain_aspect_ratio:
+            new_img = np.array(
+                pillow_img.resize(self.size, resample=self.resample))
+        else:
+            # Resize the image to the requested size in one of the dimensions.
+            # We then create a black image of the requested size and paste the
+            # resized image into the middle of this new image. The effect is
+            # that there is a black border on the top and bottom or the left
+            # and right of the resized image.
+            orig_size = pillow_img.size
+            ratio = max(self.size) / max(orig_size)
+            inter_size = (np.array(orig_size) * ratio).astype(np.int32)
+            inter_img = pillow_img.resize(inter_size, resample=self.resample)
+            new_img = Image.new('RGB', self.size)
+            upper_left = (
+                (self.size[0] - inter_size[0]) // 2,
+                (self.size[1] - inter_size[1]) // 2)
+            new_img.paste(inter_img, box=upper_left)
+            new_img = np.array(new_img)
+
+        return ImageStim(stim.filename, data=new_img)
+
+
 class PillowImageFilter(ImageFilter):
 
     ''' Uses the ImageFilter module from PIL to run a pre-defined image enhancement