+++ pliers/tests/extractors/test_text_extractors.py
@@ -76,6 +76,25 @@ def test_predefined_dictionary_extractor():
     assert np.isclose(result['aoa_Freq_pm'][0], 10.313725, 1e-5)
 
 
+def test_predefined_dictionary_retrieval():
+    variables = [
+        'affect/D.Mean.H',
+        'concreteness/SUBTLEX',
+        'subtlexusfrequency/Zipf-value',
+        'calgarysemanticdecision/RTclean_mean',
+        'massiveauditorylexicaldecision/PhonLev'
+    ]
+    stim = TextStim(text='perhaps')
+    td = PredefinedDictionaryExtractor(variables)
+    result = td.transform(stim).to_df().iloc[0]
+    assert np.isnan(result['affect_D.Mean.H'])
+    assert result['concreteness_SUBTLEX'] == 6939
+    assert result['calgarysemanticdecision_RTclean_mean'] == 954.48
+    assert np.isclose(result['subtlexusfrequency_Zipf-value'], 5.1331936)
+    assert np.isclose(result['massiveauditorylexicaldecision_PhonLev'],
+                      6.65101626)
+
+
 def test_part_of_speech_extractor():
     import nltk
     nltk.download('tagsets')
@@ -159,15 +178,16 @@ def test_vader_sentiment_extractor():
     assert result2['sentiment_neu'][0] == 0.248
     assert result2['sentiment_compound'][0] == 0.8439
 
-    
+
 def test_spacy_token_extractor():
+    pytest.importorskip('spacy')
     stim = TextStim(text='This is a test.')
     ext = SpaCyExtractor(extractor_type='token')
     assert ext.model is not None
-    
+
     ext2 = SpaCyExtractor(model='en_core_web_sm')
     assert isinstance(ext2.model, spacy.lang.en.English)
-    
+
     result = ext.transform(stim).to_df()
     assert result['text'][0] == 'This'
     assert result['lemma_'][0].lower() == 'this'
@@ -181,7 +201,7 @@ def test_spacy_token_extractor():
     assert result['is_ascii'][0] == 'True'
     assert result['is_digit'][0] == 'False'
     assert result['sentiment'][0] == '0.0'
-    
+
     assert result['text'][1] == 'is'
     assert result['lemma_'][1].lower() == 'be'
     assert result['pos_'][1] == 'VERB'
@@ -194,7 +214,7 @@ def test_spacy_token_extractor():
     assert result['is_ascii'][1] == 'True'
     assert result['is_digit'][1] == 'False'
     assert result['sentiment'][1] == '0.0'
-    
+
     assert result['text'][2] == 'a'
     assert result['lemma_'][2].lower() == 'a'
     assert result['pos_'][2] == 'DET'
@@ -207,7 +227,7 @@ def test_spacy_token_extractor():
     assert result['is_ascii'][2] == 'True'
     assert result['is_digit'][2] == 'False'
     assert result['sentiment'][2] == '0.0'
-    
+
     assert result['text'][3] == 'test'
     assert result['lemma_'][3].lower() == 'test'
     assert result['pos_'][3] == 'NOUN'
@@ -219,21 +239,22 @@ def test_spacy_token_extractor():
     assert result['is_punct'][3] == 'False'
     assert result['is_ascii'][3] == 'True'
     assert result['is_digit'][3] == 'False'
-    assert result['sentiment'][3] == '0.0'    
+    assert result['sentiment'][3] == '0.0'
 
 
 def test_spacy_doc_extractor():
+    pytest.importorskip('spacy')
     stim2 = TextStim(text='This is a test. And we are testing again. This '
                      'should be quite interesting. Tests are totally fun.')
     ext = SpaCyExtractor(extractor_type='doc')
     assert ext.model is not None
-    
+
     result = ext.transform(stim2).to_df()
     assert result['text'][0]=='This is a test. '
     assert result['is_parsed'][0]
     assert result['is_tagged'][0]
     assert result['is_sentenced'][0]
-    
+
     assert result['text'][3]=='Tests are totally fun.'
     assert result['is_parsed'][3]
     assert result['is_tagged'][3]